requirements.txt
django
mysqlclient
mysql-connector-python
google-generativeai
markdown
python-dotenv